Agregar y editar funcional

+++ app/src/main/java/com/rsamqui/bakingbills/bd/entidades/IngredienteEntity.kt
@@ -14,7 +14,7 @@ data class IngredienteEntity(
     @ColumnInfo(name = "nombre")
     val nombre: String,
     @ColumnInfo(name = "cantidad")
-    val cantidad: Int,
+    val cantidad: Double,
     @ColumnInfo(name = "precio")
     val precio: Double,
