+++ app/src/main/java/com/rsamqui/bakingbills/fragments/agregar/addIngredientsFragment.kt
@@ -28,7 +28,7 @@ class addIngredientsFragment : Fragment() {
 
     private fun guardarRegistro() {
         val name = fBinding.etIngrediente.text.toString()
-        val quantity = fBinding.etCantidad.text.toString().toInt()
+        val quantity = fBinding.etCantidad.text.toString().toDouble()
         val price = fBinding.etPrecio.text.toString().toDouble()
         val ingrediente = IngredienteEntity(0, name, quantity, price)
         viewModel.agregarIngrediente(ingrediente)
